\encoding{UTF-8}
\name{SEMSample}
\docType{data}
\alias{SEMSample}
\title{
Sample Data for SEM
}
\description{
Sample data set of repeated measurements by three observers in 20 patients.
}
\usage{
data(SEMSample)
}
\format{A data.frame containing 120 observations of 4 variables.}

\source{
    Zoran B. Popović, James D. Thomas (2017) Assessing observer variability: a user’s guide, Cardiovascular Diagnosis and Therapy, 7(3): 317-324, DOI: 10.21037/cdt.2017.03.12
}

\examples{
# load the dataset
data(aSAH)
}

\keyword{datasets}
